Creating ProvisionAwareTrait and adding it to Command classes.
