+++ src/Application.php
@@ -7,6 +7,7 @@ use Aegir\Provision\Command\ServicesCommand;
 use Aegir\Provision\Command\ShellCommand;
 use Aegir\Provision\Command\StatusCommand;
 use Aegir\Provision\Command\VerifyCommand;
+use Aegir\Provision\Common\ProvisionAwareTrait;
 use Aegir\Provision\Console\Config;
 use Aegir\Provision\Console\ConsoleOutput;
 use Drupal\Console\Core\Style\DrupalStyle;
@@ -43,21 +44,8 @@ class Application extends BaseApplication
      * @var string
      */
     const DEFAULT_TIMEZONE = 'America/New_York';
-    
-    /**
-     * @var LoggerInterface
-     */
-    public $logger;
-    
-    /**
-     * @var DrupalStyle
-     */
-    public $io;
-    
-    /**
-     * @var \Aegir\Provision\Provision
-     */
-    public $provision;
+
+    use ProvisionAwareTrait;
     
     /**
      * @var ConsoleOutput
@@ -72,7 +60,7 @@ class Application extends BaseApplication
      *
      * @throws \Exception
      */
-    public function __construct($name = 'UNKNOWN', $version = 'UNKNOWN', Provision $provision = NULL)
+    public function __construct($name = 'UNKNOWN', $version = 'UNKNOWN')
     {
         // If no timezone is set, set Default.
         if (empty(ini_get('date.timezone'))) {
@@ -87,7 +75,6 @@ class Application extends BaseApplication
 //            throw new \Exception($e->getMessage());
 //        }
 
-        $this->provision = $provision;
         parent::__construct($name, $version);
     }
     
@@ -118,7 +105,7 @@ class Application extends BaseApplication
      */
     public function getConfig()
     {
-        return $this->provision->getConfig();
+        return $this->getProvision()->getConfig();
     }
 
     /**
@@ -128,11 +115,11 @@ class Application extends BaseApplication
     {
         $commands[] = new HelpCommand();
         $commands[] = new ListCommand();
-        $commands[] = new SaveCommand();
-        $commands[] = new ServicesCommand();
+        $commands[] = new SaveCommand($this->getProvision());
+        $commands[] = new ServicesCommand($this->getProvision());
 //        $commands[] = new ShellCommand();
-        $commands[] = new StatusCommand();
-        $commands[] = new VerifyCommand();
+        $commands[] = new StatusCommand($this->getProvision());
+        $commands[] = new VerifyCommand($this->getProvision());
 
         return $commands;
     }
@@ -155,146 +142,5 @@ class Application extends BaseApplication
         );
 
         return $inputDefinition;
-    }
-    
-    /**
-     * Load all contexts into Context objects.
-     *
-     * @return array
-     */
-    static function getAllContexts($name = '', $application = NULL) {
-        $contexts = [];
-        $config = new Config();
-
-        $context_files = [];
-        $finder = new \Symfony\Component\Finder\Finder();
-        $finder->files()->name('*' . $name . '.yml')->in($config->get('config_path') . '/provision');
-        foreach ($finder as $file) {
-            list($context_type, $context_name) = explode('.', $file->getFilename());
-            $context_files[$context_name] = [
-                'name' => $context_name,
-                'type' => $context_type,
-                'file' => $file,
-            ];
-        }
-
-        foreach ($context_files as $context) {
-            $class = Context::getClassName($context['type']);
-            $contexts[$context['name']] = new $class($context['name'], $application);
-        }
-
-        if ($name && isset($contexts[$name])) {
-            return $contexts[$name];
-        }
-        elseif ($name && !isset($contexts[$name])) {
-            return NULL;
-        }
-        else {
-            return $contexts;
-        }
-    }
-
-    /**
-     * Load all server contexts.
-     *
-     * @param null $service
-     * @return mixed
-     * @throws \Exception
-     */
-    static public function getAllServers($service = NULL) {
-        $servers = [];
-        $context_files = self::getAllContexts();
-        if (empty($context_files)) {
-            throw new \Exception('No server contexts found. Use `provision save` to create one.');
-        }
-        foreach ($context_files as $context) {
-            if ($context->type == 'server') {
-                $servers[$context->name] = $context;
-            }
-        }
-        return $servers;
-    }
-
-    /**
-     * Get a simple array of all contexts, for use in an options list.
-     * @return array
-     */
-    public function getAllContextsOptions($type = NULL) {
-        $options = [];
-        foreach ($this->getAllContexts() as $name => $context) {
-            if ($type) {
-                if ($context->type == $type) {
-                    $options[$name] = $context->name;
-                }
-            }
-            else {
-                $options[$name] = $context->type . ' ' . $context->name;
-            }
-        }
-        return $options;
-    }
-    
-    /**
-     * Load the Aegir context with the specified name.
-     *
-     * @param $name
-     *
-     * @return \Aegir\Provision\Context
-     * @throws \Exception
-     */
-    static public function getContext($name, Application $application = NULL) {
-        if (empty($name)) {
-            throw new \Exception('Context name must not be empty.');
-        }
-        if (empty(Application::getAllContexts($name, $application))) {
-            throw new \Exception('Context not found with name: ' . $name);
-        }
-        return Application::getAllContexts($name, $application);
-    }
-
-    /**
-     * Get a simple array of all servers, optionally specifying the the service_type to filter by ("http", "db" etc.)
-     * @param string $service_type
-     * @return array
-     */
-    public function getServerOptions($service_type = '') {
-        $servers = [];
-        foreach ($this->getAllServers() as $server) {
-            if ($service_type && !empty($server->config['services'][$service_type])) {
-                $servers[$server->name] = $server->name . ': ' . $server->config['services'][$service_type]['type'];
-            }
-            elseif ($service_type == '') {
-                $servers[$server->name] = $server->name . ': ' . $server->config['services'][$service_type]['type'];
-            }
-        }
-        return $servers;
-    }
-
-    /**
-     * Check that a context type's service requirements are provided by at least 1 server.
-     *
-     * @param $type
-     * @return array
-     */
-    static function checkServiceRequirements($type) {
-        $class_name = Context::getClassName($type);
-
-        // @var $context Context
-        $service_requirements = $class_name::serviceRequirements();
-
-        $services = [];
-        foreach ($service_requirements as $service) {
-            try {
-                if (empty(Application::getAllServers($service))) {
-                    $services[$service] = 0;
-                }
-                else {
-                    $services[$service] = 1;
-                }
-            } catch (\Exception $e) {
-                $services[$service] = 0;
-            }
-        }
-        return $services;
     }
 }
