+++ src/ConfigDefinition/ContextNodeDefinition.php
@@ -3,6 +3,7 @@
 namespace Aegir\Provision\ConfigDefinition;
 
 use Aegir\Provision\Application;
+use Aegir\Provision\Provision;
 use Symfony\Component\Config\Definition\Builder\ScalarNodeDefinition;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 
@@ -52,10 +53,10 @@ class ContextNodeDefinition extends ScalarNodeDefinition
     {
         // No need to do anything else.
         // If there is no context named $value, getContext() throws an exception for us.
-        Application::getContext($value);
+        Provision::getContext($value);
 
         // If context_type is specified, Validate that the desired context is the right type.
-        if ($this->getNode()->getAttribute('context_type') && Application::getContext($value)->type != $this->getNode()->getAttribute('context_type')) {
+        if ($this->getNode()->getAttribute('context_type') && Provision::getContext($value)->type != $this->getNode()->getAttribute('context_type')) {
             throw new InvalidConfigurationException(strtr('The context specified for !name must be type !type.', [
                 '!name' => $this->name,
                 '!type' => $this->getNode()->getAttribute('context_type'),
@@ -70,7 +71,7 @@ class ContextNodeDefinition extends ScalarNodeDefinition
                 $path[2]
             ;
             
-            Application::getContext($value)->getService($service);
+            Provision::getContext($value)->getService($service);
         }
         return $value;
     }
