+++ src/Context/SiteContext.php
@@ -4,6 +4,7 @@ namespace Aegir\Provision\Context;
 
 use Aegir\Provision\Application;
 use Aegir\Provision\ContextSubscriber;
+use Aegir\Provision\Provision;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
 /**
@@ -42,7 +43,7 @@ class SiteContext extends ContextSubscriber implements ConfigurationInterface
         // There is no need to check if the property exists because the config system does that.
 //        $this->db_server = $application->getContext($this->properties['db_server']);
 
-        $this->platform = Application::getContext($this->properties['platform'], $application);
+        $this->platform = Provision::getContext($this->properties['platform'], $application);
 
 
         // Add platform http service subscription.
