+++ src/Provision.php
@@ -21,6 +21,7 @@ use Robo\Common\IO;
 use Robo\Contract\BuilderAwareInterface;
 use Robo\Contract\ConfigAwareInterface;
 use Robo\Contract\IOAwareInterface;
+use Robo\Log\RoboLogger;
 use Robo\Robo;
 use Robo\Runner as RoboRunner;
 use Symfony\Component\Console\Input\InputInterface;
@@ -75,7 +76,9 @@ class Provision implements ConfigAwareInterface, ContainerAwareInterface, Logger
         $this->setOutput($output);
         
         // Create Application.
-        $application = new Application(self::APPLICATION_NAME, self::VERSION, $this);
+        $application = new Application(self::APPLICATION_NAME, self::VERSION);
+        $application->setProvision($this);
+        
 //        $application->setConfig($consoleConfig);
         
         // Create and configure container.
@@ -92,6 +95,7 @@ class Provision implements ConfigAwareInterface, ContainerAwareInterface, Logger
         $this->runner->setSelfUpdateRepository(self::REPOSITORY);
     
         $this->setBuilder($container->get('builder'));
+        $this->setLogger($container->get('logger'));
     }
     
     public function run(InputInterface $input, OutputInterface $output) {
@@ -136,5 +140,147 @@ class Provision implements ConfigAwareInterface, ContainerAwareInterface, Logger
     public function getInput()
     {
         return $this->input();
+    }
+    
+    
+    /**
+     * Load all contexts into Context objects.
+     *
+     * @return array
+     */
+    static function getAllContexts($name = '', $application = NULL) {
+        $contexts = [];
+        $config = new Config();
+        
+        $context_files = [];
+        $finder = new \Symfony\Component\Finder\Finder();
+        $finder->files()->name('*' . $name . '.yml')->in($config->get('config_path') . '/provision');
+        foreach ($finder as $file) {
+            list($context_type, $context_name) = explode('.', $file->getFilename());
+            $context_files[$context_name] = [
+                'name' => $context_name,
+                'type' => $context_type,
+                'file' => $file,
+            ];
+        }
+        
+        foreach ($context_files as $context) {
+            $class = Context::getClassName($context['type']);
+            $contexts[$context['name']] = new $class($context['name'], $application);
+        }
+        
+        if ($name && isset($contexts[$name])) {
+            return $contexts[$name];
+        }
+        elseif ($name && !isset($contexts[$name])) {
+            return NULL;
+        }
+        else {
+            return $contexts;
+        }
+    }
+    
+    /**
+     * Load all server contexts.
+     *
+     * @param null $service
+     * @return mixed
+     * @throws \Exception
+     */
+    static public function getAllServers($service = NULL) {
+        $servers = [];
+        $context_files = self::getAllContexts();
+        if (empty($context_files)) {
+            throw new \Exception('No server contexts found. Use `provision save` to create one.');
+        }
+        foreach ($context_files as $context) {
+            if ($context->type == 'server') {
+                $servers[$context->name] = $context;
+            }
+        }
+        return $servers;
+    }
+    
+    /**
+     * Get a simple array of all contexts, for use in an options list.
+     * @return array
+     */
+    public function getAllContextsOptions($type = NULL) {
+        $options = [];
+        foreach ($this->getAllContexts() as $name => $context) {
+            if ($type) {
+                if ($context->type == $type) {
+                    $options[$name] = $context->name;
+                }
+            }
+            else {
+                $options[$name] = $context->type . ' ' . $context->name;
+            }
+        }
+        return $options;
+    }
+    
+    /**
+     * Load the Aegir context with the specified name.
+     *
+     * @param $name
+     *
+     * @return \Aegir\Provision\Context
+     * @throws \Exception
+     */
+    static public function getContext($name, Application $application = NULL) {
+        if (empty($name)) {
+            throw new \Exception('Context name must not be empty.');
+        }
+        if (empty(Provision::getAllContexts($name, $application))) {
+            throw new \Exception('Context not found with name: ' . $name);
+        }
+        return Provision::getAllContexts($name, $application);
+    }
+    
+    /**
+     * Get a simple array of all servers, optionally specifying the the service_type to filter by ("http", "db" etc.)
+     * @param string $service_type
+     * @return array
+     */
+    public function getServerOptions($service_type = '') {
+        $servers = [];
+        foreach ($this->getAllServers() as $server) {
+            if ($service_type && !empty($server->config['services'][$service_type])) {
+                $servers[$server->name] = $server->name . ': ' . $server->config['services'][$service_type]['type'];
+            }
+            elseif ($service_type == '') {
+                $servers[$server->name] = $server->name . ': ' . $server->config['services'][$service_type]['type'];
+            }
+        }
+        return $servers;
+    }
+    
+    /**
+     * Check that a context type's service requirements are provided by at least 1 server.
+     *
+     * @param $type
+     * @return array
+     */
+    static function checkServiceRequirements($type) {
+        $class_name = Context::getClassName($type);
+        
+        // @var $context Context
+        $service_requirements = $class_name::serviceRequirements();
+        
+        $services = [];
+        foreach ($service_requirements as $service) {
+            try {
+                if (empty(Provision::getAllServers($service))) {
+                    $services[$service] = 0;
+                }
+                else {
+                    $services[$service] = 1;
+                }
+            } catch (\Exception $e) {
+                $services[$service] = 0;
+            }
+        }
+        return $services;
     }
 }
